requirements.txt
dj-database-url==0.3.0
Django==1.9.1
django-allauth==0.24.1
django-appconf==1.0
django-crispy-forms==1.6.0
django-imagekit==3.3
django-likes==0.2
django-secretballot==0.5.0
django-taggit==0.18.0
gunicorn==19.4.5
oauthlib==1.0.3
pilkit==1.1.13
Pillow==2.9.0
psycopg2==2.6.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
wheel==0.24.0
